requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
einops==0.3.0
idna==2.10
joblib==0.16.0
jsonpatch==1.26
jsonpointer==2.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.0
opencv-python==4.5.1.48
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==19.0.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torchfile==0.1.0
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.9
visdom==0.1.8.9
websocket-client==0.57.0
